Measure the kernels against the bf16 transformer, not against ComfyUI
tools/quality_vs_bf16.py shares the noise, text states, schedule and VAE between
a Torch bf16 arm on the release checkpoint and the Loom W8A8 kernels through the
C ABI; three stages, one process each, because HRX and Torch cannot both take
the GPU while another job is on the box. Turbo at 1024^2, eight steps, seed 0:
latent PSNR 24.57 dB, image 33.67 dB, and 21.62 / 31.57 with the int4-QK
attention, so the fp16 default stands on the model's own metric.

+++ host/native_pipeline.cpp
@@ -276,7 +276,7 @@ int generate(krea2_pipeline *p, const char *prompt, const char *negative,
       uncond = p->models.text_fusion(p->models.encode(
           p->models.tokenizer.prompt(negative ? negative : "")));
     timing.mark("encode and text fusion");
-    const float mu = p->distilled ? 1.15f : dynamic_mu(w / 16 * (h / 16));
+    const double mu = p->distilled ? 1.15 : dynamic_mu(w / 16 * (h / 16));
     for (int step = 0; step < steps; ++step) {
       float sigma = scheduler_sigma(step, steps, mu),
             next = scheduler_sigma(step + 1, steps, mu);
